+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Add Recall Knowledge as optional Pathfinder Saving Throw
 + Fix Bug when quickly editing maxHP and HP
 + Refactor Party System to support two different OBR rooms loading two different parties. 
++ Add Spell Cast Type and Spell Level to PF Spells
 
 # 3.7.2